// frontend/src/app/model/hotel-package.ts
export class HotelPackage {
    id: string;
    hotelId: string;
    name: string;
    description: string;
    price: number;
    roomType: string;
    bedType: string;
    maxAdults: string;
    maxChildren: string;
    insertUserId: string;
    updateUserId: string;
    insertDate: Date;
    updateDate: Date;
    status: number;
    sortOrder: number;
}